+++ pokedex/src/pages/DetailPage.js
@@ -12,22 +12,6 @@ const DetailPage = () => {
         getPokemonDetail()
     }, [])
 
-    const getPokemonDetail = () => {
-        const urlDetailPokemon = 'https://pokeapi.co/api/v2/pokemon/:name/'
-        axios
-        .get(urlDetailPokemon)
-        .then((response) => {
-            console.log(response.data)
-            // console.log(response.data.types)
-            // console.log(response.data.stats)
-            // console.log(response.data.moves)
-            // console.log(response.data.sprites)
-        })
-        .catch((error) => {
-            console.log(error)
-        })
-    }
-
     return (
         <div>
             <p>Detail Page</p>
